+++ seatunnel-engine/seatunnel-engine-common/src/main/java/org/apache/seatunnel/engine/common/utils/serializer/ProtoStuffSerializer.java
@@ -18,7 +18,7 @@
  *
  */
 
-package org.apache.seatunnel.common.serializer;
+package org.apache.seatunnel.engine.common.utils.serializer;
 
 import io.protostuff.LinkedBuffer;
 import io.protostuff.ProtostuffIOUtil;
