+++ seatunnel-engine/seatunnel-engine-common/src/main/java/org/apache/seatunnel/engine/common/utils/serializer/Serializer.java
@@ -18,7 +18,7 @@
  *
  */
 
-package org.apache.seatunnel.common.serializer;
+package org.apache.seatunnel.engine.common.utils.serializer;
 
 import java.io.IOException;
