+++ seatunnel-engine/seatunnel-engine-storage/imap-storage-file/pom.xml
@@ -33,14 +33,8 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.seatunnel</groupId>
-            <artifactId>seatunnel-common</artifactId>
+            <artifactId>seatunnel-engine-common</artifactId>
             <version>${project.version}</version>
-            <exclusions>
-                <exclusion>
-                    <artifactId>guava</artifactId>
-                    <groupId>com.google.guava</groupId>
-                </exclusion>
-            </exclusions>
         </dependency>
         <dependency>
             <groupId>org.apache.seatunnel</groupId>
